Check for .NET 5 earlier
